feat: Add font configuration for Inter and Montserrat_Alternates
This commit adds a new file, `fonts.ts`, which contains the configuration for the Inter and Montserrat_Alternates fonts. The `inter` font is imported from the `next/font/google` package and is configured with the subset "latin". The `titleFont` is also imported from the same package and is configured with the subsets "latin" and the weights "500" and "700".