feat: Include TopMenu component in ShopLayout
This commit adds the TopMenu component to the ShopLayout in the src/components/index.ts file. The TopMenu component is now exported from the components module, allowing it to be used in other parts of the application. This change is necessary to provide a consistent navigation experience throughout the shop pages.