feat: Update npm dependencies
This commit updates the npm dependencies in the package.json and package-lock.json files. It adds the "clsx" and "zustand" packages to the dependencies, and updates the versions of "next", "react", "react-dom", and "react-icons". This change enhances the functionality and performance of the application by incorporating the latest stable versions of these dependencies.