feat: Add openMenu function to TopMenu component
This commit adds the openMenu function to the TopMenu component in the src/components/ui/top-menu/TopMenu.tsx file. The openMenu function is called when the "Menu" button is clicked and it triggers the opening of the side menu. This change enhances the functionality of the application by providing a way for users to easily access the side menu.