refactor: Update CartPage checkout button styling
This commit updates the styling of the checkout button on the CartPage. The previous styling used the "flex" class, which caused the button to be centered horizontally. However, this resulted in the button being misaligned with the rest of the content. To fix this, the "justify-center" class is added to the button, ensuring that it is properly aligned within its container. This improvement enhances the user experience by providing a visually consistent layout on the CartPage.