refactor: Update seed-database.ts for database seeding
This commit refactors the "seed-database.ts" file to improve the process of seeding the database with initial data. It removes the commented out code for deleting data and instead uses the Prisma deleteMany() method to delete existing records from the "productImage", "product", and "category" tables. This change ensures that the database is properly cleared before adding new data. Additionally, it introduces a more efficient approach for adding categories and products by using the Prisma createMany() method. These improvements enhance the reliability and performance of the database seeding process.