refactor: Update shop page to use paginated product data
The code changes in this commit update the shop page to use paginated product data instead of the initial data from the seed file. The `getPaginatedProductsWithImages` function is now imported and used to fetch the products based on the provided search parameters. If there are no products returned, the page is redirected to the home page. This refactor improves the functionality of the shop page by implementing pagination and dynamically fetching the product data.