refactor: Update .prettierrc configuration
This commit adds a new .prettierrc file with updated configuration settings. The printWidth is set to 100, useTabs is set to false, tabWidth is set to 2, trailingComma is set to "es5", semi is set to false, singleQuote is set to true, bracketSpacing is set to true, arrowParens is set to "always", jsxSingleQuote is set to false, bracketSameLine is set to true, endOfLine is set to "lf", and singleAttributePerLine is set to false. These changes ensure consistent code formatting and improve code readability.