feat: Add generatePaginationNumbers utility function
This commit adds the `generatePaginationNumbers` utility function to the `src/utils/generatePaginationNumbers.ts` file. The function takes the current page number and the total number of pages as parameters and returns an array of page numbers to be used for pagination. It handles different scenarios such as when there are less than 7 pages, when the current page is less than 3, when the current page is greater than totalPages - 2, and when the current page is between 3 and totalPages - 2. This utility function enhances the pagination functionality of the application by generating the appropriate page numbers for navigation.