refactor: Update QuantitySelector component to handle onQuantityChange prop
This commit refactors the QuantitySelector component in the src/components/product/quantity-selector/QuantitySelector.tsx file to handle the onQuantityChange prop. The onQuantityChange prop is now passed as a prop to the QuantitySelector component, allowing for a callback function to be executed when the quantity is changed. This refactor improves the flexibility and reusability of the QuantitySelector component.