feat: Add getTotalItems function to cart store
This commit adds the getTotalItems function to the cart store in the cart-store.ts file. The function calculates the total number of items in the cart by summing up the quantities of all cart items. This feature enhances the cart functionality by providing an easy way to retrieve the total number of items in the cart.