feat: Add currencyFormater utility function
The currencyFormater utility function has been added to the src/utils/currencyFormater.ts file. This function formats a given number as currency using the 'en-US' locale and the Euro currency. It ensures that the value is displayed with a minimum of 2 decimal places and a maximum of 2 decimal places. This feature enhances the application by providing a consistent and user-friendly way to format currency values.