feat: Export currencyFormater utility function from utils
The currencyFormater utility function has been added to the `src/utils` directory and exported from the `index.ts` file. This function is used to format a given number as currency using the 'en-US' locale and the Euro currency. It ensures that the value is displayed with a minimum of 2 decimal places and a maximum of 2 decimal places. This change enhances the application by providing a consistent and user-friendly way to format currency values.