refactor: Update AddressForm component to save user address to database
This code change updates the AddressForm component to include logic for saving the user's address to the database. When the form is submitted and the "rememberAddress" checkbox is checked, the setUserAddress action is called to save the address data. This enhancement improves the functionality of the form by persisting the user's address for future use.