refactor: Delete user address in deleteUserAddress function
This code change adds a new function, deleteUserAddress, which deletes a user's address from the database. The function takes a userId as a parameter and uses Prisma to delete the corresponding user address. If the deletion is successful, the function returns a success message along with the deleted address. If an error occurs during the deletion process, the function returns an error message. This enhancement improves the functionality of the code by providing a way to delete user addresses from the database.